+++ src/utils/utils.js
@@ -1,0 +1,5 @@
+/**
+ * Utils namespace
+ * @type {Object}
+ */
+mCAP.Utils = {};